@@ -3,6 +3,10 @@
 	"version": "0.2.0",
 	"description": "",
 	"main": "src/index.js",
+	"type": "module",
+	"engines": {
+		"node": ">=20.13.0"
+	},
 	"scripts": {
 		"build": "tailwind -i ./src/client/site.css -o ./public/assets/site.css",
 		"dev": "nodemon --watch src --watch public -e js,html,css --exec \"npm run build && npm run start\"",
@@ -22,16 +26,16 @@
 		"url": "https://github.com/reverentgeek/node-bacon-generator.git"
 	},
 	"devDependencies": {
-		"eslint": "^8.42.0",
-		"eslint-config-reverentgeek": "^4.0.2",
+		"eslint": "^9.14.0",
+		"eslint-config-reverentgeek": "^5.1.0",
 		"nodemon": "^3.1.7",
 		"tailwindcss": "^3.3.2"
 	},
 	"dependencies": {
-		"@fastify/static": "^6.10.2",
+		"@fastify/static": "^8.0.2",
 		"@fontsource/montserrat": "^5.0.3",
 		"dotenv": "^16.2.0",
-		"fastify": "^4.18.0",
+		"fastify": "^5.1.0",
 		"fs-extra": "^11.1.1",
 		"vue": "^3.3.4",
 		"vue3-clipboard": "^1.0.0"
