@@ -1,23 +1,20 @@
-"use strict";
 
-const path = require( "path" );
-const api = require( "./api" );
-const staticRoutes = require( "./static" );
+import { register as api } from "./api.js";
+import { register as staticRoutes } from "./static.js";
 
-function register( server ) {
-	api.register( server );
-	staticRoutes.register( server );
+import { join } from "node:path";
+
+const __dirname = import.meta.dirname;
+
+export function register( server ) {
+	api( server );
+	staticRoutes( server );
 
 	// Path to views folder
-	const views = path.join( __dirname, "..", "views" );
+	const views = join( __dirname, "..", "views" );
 
 	// Declare default route
-	server.get( "/", async ( request, reply ) => {
+	server.get( "/", ( _request, reply ) => {
 		return reply.sendFile( "index.html", views );
 	} );
-
 }
-
-module.exports = {
-	register
-};
