@@ -1,17 +1,17 @@
-"use strict";
+import fastifyStatic from "@fastify/static";
+import { join } from "node:path";
 
-const fastifyStatic = require( "@fastify/static" );
-const path = require( "path" );
+const __dirname = import.meta.dirname;
 
-function register( server ) {
+export function register( server ) {
 	// Register the static file plugin
-	const publicPath = path.join( __dirname, "..", "..", "public" );
+	const publicPath = join( __dirname, "..", "..", "public" );
 	server.register( fastifyStatic, {
 		root: publicPath
 	} );
 
 	// Serve fonts
-	const fontPath = path.join( __dirname, "..", "..", "node_modules", "@fontsource", "montserrat", "files" );
+	const fontPath = join( __dirname, "..", "..", "node_modules", "@fontsource", "montserrat", "files" );
 	server.register( fastifyStatic, {
 		root: fontPath,
 		prefix: "/assets/files/",
@@ -20,7 +20,7 @@ function register( server ) {
 	} );
 
 	// Serve vue.js static files
-	const vuePath = path.join( __dirname, "..", "..", "node_modules", "vue", "dist" );
+	const vuePath = join( __dirname, "..", "..", "node_modules", "vue", "dist" );
 	server.register( fastifyStatic, {
 		root: vuePath,
 		prefix: "/vue/",
@@ -29,7 +29,7 @@ function register( server ) {
 	} );
 
 	// Serve vue clipboard static files
-	const vueClipboardPath = path.join( __dirname, "..", "..", "node_modules", "vue3-clipboard", "dist" );
+	const vueClipboardPath = join( __dirname, "..", "..", "node_modules", "vue3-clipboard", "dist" );
 	server.register( fastifyStatic, {
 		root: vueClipboardPath,
 		prefix: "/vueClip/",
@@ -37,7 +37,3 @@ function register( server ) {
 	// reply decorator has been added by the first plugin registration
 	} );
 }
-
-module.exports = {
-	register
-};
